requirements.txt
Django>=5.0,<6.0
djangorestframework
django-cors-headers
gunicorn
whitenoise
psycopg2-binary
dj-database-url
